+++ Sources/Protocols.swift
@@ -9,7 +9,11 @@
 import UIKit
 
 /// A block of code
-public typealias CompletionBlock = () -> Void
+public typealias CodeBlock = () -> Void
+
+/// A block of code
+@available(*, deprecated, message: "Deprecated in version 1.2. Use CodeBlock instead", renamed: "CodeBlock")
+public typealias CompletionBlock = CodeBlock
 
 
 // MARK: Protocols
@@ -35,7 +39,7 @@ public protocol ModuleProtocol: class {
 		- style: The modal transition style to use
 		- completion: The code block to call when the transition completes
 	*/
-	func present(from fromVC: UIViewController?, style: UIModalTransitionStyle, completion: CompletionBlock?)
+	func present(from fromVC: UIViewController?, style: UIModalTransitionStyle, completion: CodeBlock?)
 	
 	/** Pushes the View Controller owned by the Module
 	- parameters:
@@ -66,13 +70,13 @@ public protocol PresenterRouterProtocol: class {
 		- view: The view controller to present
 		- completion: The code block to call when the transition completes
 	*/
-	func present(_ view: UIViewController, completion: CompletionBlock?)
+	func present(_ view: UIViewController, completion: CodeBlock?)
 	
 	/** Dismisses a Module that was presented modally
 	- parameters:
 		- completion: The code block to call when the transition completes
 	*/
-	func dismiss(completion: CompletionBlock?)
+	func dismiss(completion: CodeBlock?)
 	
 	/** Pops a Module that was pushed
 	- returns: Should return the view controller that was popped
@@ -113,5 +117,5 @@ public protocol StoryboardIdentifiable: class {
 
 /** This Protocol is used to initialize a View Controller via. Storyboard
 */
-@available(*, deprecated, message: "Deprecated in version `1.1`. Use StoryboardIdentifiable instead", renamed: "StoryboardIdentifiable")
+@available(*, deprecated, message: "Deprecated in version 1.1. Use StoryboardIdentifiable instead", renamed: "StoryboardIdentifiable")
 public typealias ViewStoryboardProtocol = StoryboardIdentifiable