+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### Updated
 - `viewLoaded()` no longer optional for Presenters conforming to `ViewPresenterProtocol`
+- `CompletionBlock` was renamed `CodeBlock`
 
 #### Deprecated
 - `ViewPresenterProtocol.viewAppearing()`